requirements.txt
flask
python-docx
PyPDF2
sentence_transformers
regex
nltk
faiss